consolidated dsfields into a single plugin
